Fixed Zone selector in Record bulk edit view
fixes #49

The problem is that the "data-url" attribute of the APISelect widget really needs
to be an URL, not the identifier created by the ViewSet. So a call to "reverse_lazy"
is necessary to resolve that issue.

"reverse" does not work as calling it leads to a recursive include loop.