Added very sketchy tests for Auto SOA SERIAL
The problem with testing this feature is that the serial numbers it creates are
time-based, and so it would be very time-consuming to create meaningful tests, as
lots of delays would have to be implemented to make time advance in the test
environment.

One way to fix this would be to run tests asynchronsously and keep the delays
short, say in between 0 and 20 seconds. Asynchronous execution with e.g. pytest-async
would have the advantage of only having to wait for the duration of the slowest test.

Another way could be to change the "last_updated" field for certain objects in the
database. I'm not really sure if that would be such a great idea, though.

All in all, a tricky one.