Make record TTL field optional
This change is required to make it possible to have resource record TTL follow the zone
default TTL value. The TTL field for records is now optional.

=== BREAKING CHANGE ===

* API can return null for record TTL

This change will likely make it necessary to update integrations such as the example zone
exporter script provided by the module. If an integration tool relies on the record TTL
value to be set to an integer value, getting a None value instead may have undesired results.

See the example zone exporter script for a way to solve this in Ansible/Jinja 2. The solution
for other integrations will depend on the mechanisms used by the integration tool. The JSON
value returned when TTL is not set for a record is "null":

      "value": "1.1.1.1",
      "status": "active",
->    "ttl": null,
      "description": "",
      "tags": [],

* Null record TTL field results in dynamic record TTL

Records created without an explicit TTL value were formerly saved with the default TTL of
the zone in their ttl attribute. When the zone's default TTL was changed, the records kept
their existing value instead of following the change.

With this new commit, records having no explicit TTL set will have their TTL calculated
from the SOA MINIMUM field of the zone (the last one in the SOA tuple). This can be overriden
by specifying a default TTL in the zone. See RFC2308 <https://www.rfc-editor.org/rfc/rfc2308>
for details.